added simple interacting with the model (colors), and started to do ray casting to get the world coordinates from the mouse click. TODO: Test for Intersection with one for the models cells.

+++ src/LayerStack.h
@@ -9,7 +9,7 @@ public:
 
   virtual void OnAttach(){};
   virtual void OnDetach(){};
-  virtual void OnUpdate(const double timeStamp){};
+  virtual void OnUpdate(const float timeStamp){};
   virtual void OnImGuiRender(){};
   virtual void OnEvent(GLCore::Event &event){};
